+++ src/main/java/ladder/domain/result/GameResult.java
@@ -1,6 +1,5 @@
 package ladder.domain.result;
 
-import ladder.domain.Records;
 import ladder.domain.participants.Participants;
 import ladder.utils.SplitUtils;
